feat: update coin-item layout
